Time: 655 ms (31.76%), Space: 78.8 MB (28.93%) - LeetHub

+++ 1461-check-if-a-string-contains-all-binary-codes-of-size-k/1461-check-if-a-string-contains-all-binary-codes-of-size-k.cpp
@@ -15,7 +15,7 @@ public:
             j++;
         }
         
-        long long num = pow(2, k);
+        int num = pow(2, k);
         
         if(mp.size()<num) return false;
         else return true;
